Potential fix for code scanning alert no. 24: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using a larger type to prevent overflow. This can be achieved by casting one of the operands to unsigned long before performing the multiplication. This way, the multiplication will be done using unsigned long, and the result will not overflow.

Specifically, we will cast jp2idx->SIZ.XTnum to unsigned long before multiplying it with jp2idx->SIZ.YTnum. This change will be made on line 539 of the file libopenjpeg-2.3.1/src/lib/openjpip/index_manager.c.

+++ libopenjpeg-2.3.1/src/lib/openjpip/index_manager.c
@@ -536,7 +536,7 @@ OPJ_BOOL set_thixdata(box_param_t *cidx_box, index_param_t *jp2idx)
     mhixseqoff = manf_box->offset + (OPJ_OFF_T)manf_box->length;
     pos = 0;
     tile_no = 0;
-    jp2idx->tileheader = (mhixbox_param_t **)opj_malloc(jp2idx->SIZ.XTnum *
+    jp2idx->tileheader = (mhixbox_param_t **)opj_malloc((unsigned long)jp2idx->SIZ.XTnum *
                          jp2idx->SIZ.YTnum * sizeof(mhixbox_param_t *));
 
     while (ptr) {
